# game-engine/openapi.yaml
openapi: 3.1.0
info:
  title: OYNA Lasers CRM API
  version: "1.0.0"
  description: |
    Backend API для интеграции с игровым движком OYNA Lasers.

    **Актуальные пути NestJS** (префикс `/api` на сервере): сводная таблица в [CRM_ENGINE_MAPPING.md](CRM_ENGINE_MAPPING.md).
    Секция `paths` ниже начинается с реализованных маршрутов; блоки `/sessions`, `/users`, `/rating` — целевой draft (ещё не все совпадают с репозиторием oyna).

    Три группы вызовов:
      1. Room Agent → Backend (`StationApiKey`) — heartbeat, очередь команд, match-reports
      2. CRM UI → Backend (JWT оператора) — game-sessions, rooms
      3. Команды CRM → агент — через очередь `GET /station/rooms/{roomId}/commands` (polling)
  contact:
    name: Azamat
    url: https://oyna.pro
servers:
  - url: http://localhost:3000/api
    description: Локальный Nest (Swagger /api/docs)
  - url: https://api.oyna.pro/api
    description: Production (при деплое с глобальным префиксом /api)
...
components:
  securitySchemes:
    StationApiKey:
      type: http
      scheme: bearer
      bearerFormat: "station_live_<hex64>"
      description: API key конкретной станции (middleware → backend)
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Пользовательский JWT (web client → backend)
    OperatorJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT оператора (админка → backend)
  parameters:
...
    StationId:
      name: stationId
      in: path
      required: true
      schema: { type: string, example: "laser-matrix-01" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "INSUFFICIENT_BALANCE"
        message:
          type: string
          example: "User has 5 minutes, but 10 required"
        details:
          type: object
          additionalProperties: true
    HeartbeatRequest:
      type: object
      required: [buildVersion, uptime]
      properties:
        buildVersion: { type: string, example: "1.0.0" }
        uptime: { type: integer, description: "seconds since start", example: 3600 }
        currentSessionId:
          type: string
          nullable: true
          example: "sess_..."
        middlewareStatus:
          type: string
          enum: [healthy, degraded, error]
          example: "healthy"
        hardwareStatus:
          type: string
          enum: [connected, disconnected, error]
          example: "connected"
    HeartbeatResponse:
      type: object
      properties:
        acknowledged: { type: boolean }
        serverTime: { type: string, format: date-time }
        needsUpdate: { type: boolean, example: false }
        pendingCommands:
          type: array
          items: { $ref: "#/components/schemas/BackendCommand" }
    BackendCommand:
      type: object
      description: "Команда от backend которую middleware должен передать UE"
      properties:
        commandId: { type: string }
        type:
          type: string
          enum: [session_created, session_start, session_pause, session_resume, session_cancel]
        payload: { type: object, additionalProperties: true }
...
paths:
...
  /station/rooms/{roomId}/heartbeat:
    post:
      tags: [station]
      summary: Heartbeat станции
      security: [{ StationApiKey: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                buildVersion: { type: string }
                uptimeSeconds: { type: integer }
                currentSessionId: { type: string, format: uuid, nullable: true }
      responses:
        "200": { description: OK }
  /station/rooms/{roomId}/commands:
    get:
      tags: [station]
      summary: Очередь команд CRM → агент (polling)
      security: [{ StationApiKey: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Массив команд с полями id, type, payload, sessionId, createdAt
  /station/rooms/{roomId}/commands/ack:
    post:
      tags: [station]
      summary: Подтвердить доставку команд
      security: [{ StationApiKey: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [commandIds]
              properties:
                commandIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  acknowledged: { type: integer }
  /station/rooms/{roomId}/sessions/{sessionId}/match-reports:
    post:
      tags: [station]
      summary: Отчёт уровня (идемпотентно)
      security: [{ StationApiKey: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [level, attemptNumber, result]
              properties:
                level: { type: integer, minimum: 1 }
                attemptNumber: { type: integer, minimum: 1 }
                durationSeconds: { type: integer }
                completedAt: { type: string }
                result: { type: object, additionalProperties: true }
      responses:
        "200": { description: OK }
  /station/rooms/{roomId}/sessions/activate-by-token:
    post:
      tags: [station]
      summary: Активировать сессию (агент; sessionToken должен принадлежать roomId)
      security: [{ StationApiKey: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionToken]
              properties:
                sessionToken: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /station/rooms/{roomId}/sessions/{sessionId}/end:
    post:
      tags: [station]
      summary: Завершить сессию от станции (те же durationSeconds + results)
      security: [{ StationApiKey: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [durationSeconds, results]
              properties:
                durationSeconds: { type: integer, minimum: 1 }
                results:
                  type: array
                  items:
                    type: object
                    required: [userId, score]
                    properties:
                      userId: { type: string, format: uuid }
                      score: { type: integer, minimum: 0 }
      responses:
        "200": { description: OK }
...
  /stations/{stationId}/heartbeat:
    post:
      tags: [stations]
      summary: Heartbeat от middleware (каждые 30 сек)
      security: [{ StationApiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/StationId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HeartbeatRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HeartbeatResponse" }
        "401":
          description: Invalid API key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ============================================================================
  # SESSIONS
  # ============================================================================
...